ModelService/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
python-multipart==0.0.20
pillow==11.3.0
numpy==2.3.2
torch==2.8.0
torchvision==0.23.0
